+++ 301/CO_SYNC.c
@@ -417,7 +417,7 @@ CO_SYNC_status_t CO_SYNC_process(CO_SYNC_t *SYNC,
                 }
 
                 if (SYNC->timer > periodTimeout) {
-                    CO_errorReport(SYNC->em, (uint8_t)CO_EM_SYNC_TIME_OUT,
+                    CO_errorReport(SYNC->em, CO_EM_SYNC_TIME_OUT,
                                    CO_EMC_COMMUNICATION, SYNC->timer);
                     SYNC->timeoutError = 2;
                 }
@@ -448,7 +448,7 @@ CO_SYNC_status_t CO_SYNC_process(CO_SYNC_t *SYNC,
 
         /* verify error from receive function */
         if (SYNC->receiveError != 0U) {
-            CO_errorReport(SYNC->em, (uint8_t)CO_EM_SYNC_LENGTH,
+            CO_errorReport(SYNC->em, CO_EM_SYNC_LENGTH,
                            CO_EMC_SYNC_DATA_LENGTH, SYNC->receiveError);
             SYNC->receiveError = 0;
         }
@@ -462,7 +462,7 @@ CO_SYNC_status_t CO_SYNC_process(CO_SYNC_t *SYNC,
 
     if (syncStatus == CO_SYNC_RX_TX) {
         if (SYNC->timeoutError == 2U) {
-            CO_errorReset(SYNC->em, (uint8_t)CO_EM_SYNC_TIME_OUT, 0);
+            CO_errorReset(SYNC->em, CO_EM_SYNC_TIME_OUT, 0);
         }
         SYNC->timeoutError = 1;
     }
